+++ Arrangement_on_surface_2/demo/Arrangement_on_surface_2_earth/Aos.h
@@ -76,9 +76,9 @@ public:
   static std::string locate_country(Arr_handle arrh, const QVector3D& point);
 
   // this will get the approximate arcs of face-edges from the arrangement
-  // NOTE: this is similar to "get_approx_arcs(KML::Placemarks&, float)" above!
+  // NOTE: this is similar to "get_approx_arcs(KML::Placemarks&, double)" above!
   static Approx_arcs get_approx_arcs_from_faces_edges(Arr_handle arrh,
-                                                      float error);
+                                                      double error);
 };
 
 #endif
